Update and add integration tests
