+++ packages/reporter/cypress/integration/test_errors_spec.js
@@ -1,131 +1,5 @@
 const { EventEmitter } = require('events')
-const _ = Cypress._
-
-const itHandlesFileOpening = (containerSelector) => {
-  beforeEach(function () {
-    cy.stub(this.runner, 'emit').callThrough()
-    this.setError(this.commandErr)
-  })
-
-  describe('when user has already set opener and opens file', function () {
-    beforeEach(function () {
-      this.editor = {}
-
-      this.runner.emit.withArgs('get:user:editor').yields({
-        preferredOpener: this.editor,
-      })
-
-      cy.contains('View stack trace').click()
-    })
-
-    it('opens in preferred opener', function () {
-      cy.get(`${containerSelector} a`).first().click().then(() => {
-        expect(this.runner.emit).to.be.calledWith('open:file', {
-          where: this.editor,
-          file: '/me/dev/my/app.js',
-          line: 2,
-          column: 7,
-        })
-      })
-    })
-  })
-
-  describe('when user has not already set opener and opens file', function () {
-    const availableEditors = [
-      { id: 'computer', name: 'On Computer', isOther: false, openerId: 'computer' },
-      { id: 'atom', name: 'Atom', isOther: false, openerId: 'atom' },
-      { id: 'vim', name: 'Vim', isOther: false, openerId: 'vim' },
-      { id: 'sublime', name: 'Sublime Text', isOther: false, openerId: 'sublime' },
-      { id: 'vscode', name: 'Visual Studio Code', isOther: false, openerId: 'vscode' },
-      { id: 'other', name: 'Other', isOther: true, openerId: '' },
-    ]
-
-    beforeEach(function () {
-      this.runner.emit.withArgs('get:user:editor').yields({ availableEditors })
-      // usual viewport of only reporter is a bit cramped for the modal
-      cy.viewport(600, 600)
-      cy.contains('View stack trace').click()
-      cy.get(`${containerSelector} a`).first().click()
-    })
-
-    it('opens modal with available editors', function () {
-      _.each(availableEditors, ({ name }) => {
-        cy.contains(name)
-      })
-
-      cy.contains('Other')
-      cy.contains('Set preference and open file')
-    })
-
-    // NOTE: this fails because mobx doesn't make the editors observable, so
-    // the changes to the path don't bubble up correctly. this only happens
-    // in the Cypress test and not when running the actual app
-    it.skip('updates "Other" path when typed into', function () {
-      cy.contains('Other').find('input[type="text"]').type('/path/to/editor')
-      .should('have.value', '/path/to/editor')
-    })
-
-    describe('when editor is not selected', function () {
-      it('disables submit button', function () {
-        cy.contains('Set preference and open file')
-        .should('have.class', 'is-disabled')
-        .click()
-
-        cy.wrap(this.runner.emit).should('not.to.be.calledWith', 'set:user:editor')
-        cy.wrap(this.runner.emit).should('not.to.be.calledWith', 'open:file')
-      })
-
-      it('shows validation message when hovering over submit button', function () {
-        cy.get('.editor-picker-modal .submit').trigger('mouseover')
-        cy.get('.cy-tooltip').should('have.text', 'Please select a preference')
-      })
-    })
-
-    describe('when Other is selected but path is not entered', function () {
-      beforeEach(function () {
-        cy.contains('Other').click()
-      })
-
-      it('disables submit button', function () {
-        cy.contains('Set preference and open file')
-        .should('have.class', 'is-disabled')
-        .click()
-
-        cy.wrap(this.runner.emit).should('not.to.be.calledWith', 'set:user:editor')
-        cy.wrap(this.runner.emit).should('not.to.be.calledWith', 'open:file')
-      })
-
-      it('shows validation message when hovering over submit button', function () {
-        cy.get('.editor-picker-modal .submit').trigger('mouseover')
-        cy.get('.cy-tooltip').should('have.text', 'Please enter the path for the "Other" editor')
-      })
-    })
-
-    describe('when editor is set', function () {
-      beforeEach(function () {
-        cy.contains('Visual Studio Code').click()
-        cy.contains('Set preference and open file').click()
-      })
-
-      it('closes modal', function () {
-        cy.contains('Set preference and open file').should('not.be.visible')
-      })
-
-      it('emits set:user:editor', function () {
-        expect(this.runner.emit).to.be.calledWith('set:user:editor', availableEditors[4])
-      })
-
-      it('opens file in selected editor', function () {
-        expect(this.runner.emit).to.be.calledWith('open:file', {
-          where: availableEditors[4],
-          file: '/me/dev/my/app.js',
-          line: 2,
-          column: 7,
-        })
-      })
-    })
-  })
-}
+const { itHandlesFileOpening } = require('../support/utils')
 
 describe('test errors', function () {
   beforeEach(function () {
@@ -353,7 +227,11 @@ describe('test errors', function () {
       cy.get('.command-wrapper').should('be.visible')
     })
 
-    itHandlesFileOpening('.runnable-err-stack-trace')
+    itHandlesFileOpening('.runnable-err-stack-trace', {
+      file: '/me/dev/my/app.js',
+      line: 2,
+      column: 7,
+    }, true)
   })
 
   describe('command error', function () {
@@ -405,9 +283,11 @@ describe('test errors', function () {
   })
 
   describe('code frames', function () {
-    it('shows code frame when included on error', function () {
+    beforeEach(function () {
       this.setError(this.commandErr)
+    })
 
+    it('shows code frame when included on error', function () {
       cy
       .get('.test-err-code-frame')
       .should('be.visible')
@@ -415,7 +295,6 @@ describe('test errors', function () {
 
     it('does not show code frame when not included on error', function () {
       this.commandErr.codeFrame = undefined
-      this.setError(this.commandErr)
 
       cy
       .get('.test-err-code-frame')
@@ -423,8 +302,6 @@ describe('test errors', function () {
     })
 
     it('use correct language class', function () {
-      this.setError(this.commandErr)
-
       cy
       .get('.test-err-code-frame pre')
       .should('have.class', 'language-javascript')
@@ -432,13 +309,15 @@ describe('test errors', function () {
 
     it('falls back to text language class', function () {
       this.commandErr.codeFrame.language = null
-      this.setError(this.commandErr)
-
       cy
       .get('.test-err-code-frame pre')
       .should('have.class', 'language-text')
     })
 
-    itHandlesFileOpening('.test-err-code-frame')
+    itHandlesFileOpening('.test-err-code-frame', {
+      file: '/me/dev/my/app.js',
+      line: 2,
+      column: 7,
+    })
   })
 })
