+++ packages/reporter/src/opener/editor-picker-modal.tsx
