+++ packages/reporter/src/opener/file-opener.tsx
