+++ packages/reporter/src/opener/opener.scss
